Uploaded the actual workfiles.
The original folder in this repo was not the code I was actually working on. It was and then I made a copy of the files and moved them into a directory that was much easier to navigate to and resumed coding in that directory without creating a git repo for it.